# Fast Hysteresis Thresholding for Canny