--- a/icarus_signal_processing/Detection/LinkDef.h
+++ b/icarus_signal_processing/Detection/LinkDef.h
@@ -14,6 +14,7 @@
 #pragma link C++ class icarus_signal_processing::Detection::MorphologicalFunctions1D+;
 #pragma link C++ class icarus_signal_processing::Detection::MorphologicalFunctions2D+;
 #pragma link C++ class icarus_signal_processing::Detection::Thresholding+;
+#pragma link C++ class icarus_signal_processing::Detection::DisjointSetForest+;
 //ADD_NEW_CLASS ... do not change this line
 #endif