This is a better implementation of double+hysteresis thresholding which is ~100x faster (at least in CPU w/o multithreading)

Old:
```
18.8 s ± 157 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

New:
```
176 ms ± 11.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Here `EdgeDetection::Canny` is modified to use the faster version.