@SFBayLaser 2021-04-15
Speed comes at a price... running at 8 GB now. But that gives incentive to get the multithreading running again once we are sorted out with what we are doing!